Conflicts:
	plugin/src/main/java/com/atlassian/plugin/connect/plugin/installer/ConnectAddonManager.java
	pom.xml

+++ plugin/src/test/java/it/com/atlassian/plugin/connect/usermanagement/jira/AbstractJiraPermissionScopeTest.java
@@ -18,6 +18,7 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import it.com.atlassian.plugin.connect.TestAuthenticator;
 import org.apache.commons.lang3.StringUtils;
+import org.junit.After;
 import org.junit.Before;
 
 import java.io.IOException;
@@ -26,6 +27,8 @@ import java.util.List;
 import static com.atlassian.plugin.connect.modules.beans.AuthenticationBean.newAuthenticationBean;
 import static com.atlassian.plugin.connect.modules.beans.ConnectAddonBean.newConnectAddonBean;
 import static com.atlassian.plugin.connect.modules.beans.LifecycleBean.newLifecycleBean;
+import static com.atlassian.plugin.connect.test.util.AddonUtil.randomPluginKey;
+import static com.atlassian.plugin.connect.test.util.AddonUtil.randomWebItemBean;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
@@ -67,35 +70,56 @@ public abstract class AbstractJiraPermissionScopeTest
     @Before
     public void setUp()
     {
+        
         ConnectAddonBean baseBean = newConnectAddonBean()
-                .withName("JIRA Project Admin Add-on")
                 .withKey(ADDON_KEY)
+                .withBaseurl(testPluginInstaller.getInternalAddonBaseUrl(ADDON_KEY))
+                .withName("JIRA Project Admin Add-on")
                 .withAuthentication(newAuthenticationBean().withType(AuthenticationType.JWT).build())
                 .withLifecycle(
                         newLifecycleBean()
                                 .withInstalled(INSTALLED)
                                 .build()
                 )
-                .withBaseurl(testPluginInstaller.getInternalAddonBaseUrl(ADDON_KEY))
+                .withModule("webItems", randomWebItemBean())
                 .build();
 
         this.adminAddOn = newConnectAddonBean(baseBean)
+                
                 .withScopes(Sets.newHashSet(ScopeName.ADMIN))
                 .build();
+
         this.projectAdminAddOn = newConnectAddonBean(baseBean)
                 .withScopes(Sets.newHashSet(ScopeName.PROJECT_ADMIN))
                 .build();
+
         this.deleteAddOn = newConnectAddonBean(baseBean)
                 .withScopes(Sets.newHashSet(ScopeName.DELETE))
                 .build();
+
         this.writeAddOn = newConnectAddonBean(baseBean)
                 .withScopes(Sets.newHashSet(ScopeName.WRITE))
                 .build();
+
         this.readAddOn = newConnectAddonBean(baseBean)
                 .withScopes(Sets.newHashSet(ScopeName.READ))
                 .build();
 
         testAuthenticator.authenticateUser("admin");
+    }
+    
+    @After
+    public void resetBeans()
+    {
+        this.adminAddOn = null;
+
+        this.projectAdminAddOn = null;
+
+        this.deleteAddOn = null;
+
+        this.writeAddOn = null;
+
+        this.readAddOn = null;
     }
 
     public ConnectAddonBean getAdminAddOn()
@@ -209,7 +233,7 @@ public abstract class AbstractJiraPermissionScopeTest
             }
             plugin = testPluginInstaller.installAddon(to);
 
-            List<String> projectAdminErrors = permissionsForAllProjects(permission, permissionMustExist);
+            List<String> projectAdminErrors = permissionsForAllProjects(permission, permissionMustExist, plugin);
             assertTrue(StringUtils.join(projectAdminErrors, '\n'), projectAdminErrors.isEmpty());
         }
         finally
@@ -218,7 +242,7 @@ public abstract class AbstractJiraPermissionScopeTest
         }
     }
 
-    protected List<String> permissionsForAllProjects(Permission permission, boolean permissionMustExist) throws ConnectAddOnUserInitException
+    protected List<String> permissionsForAllProjects(Permission permission, boolean permissionMustExist, Plugin plugin) throws ConnectAddOnUserInitException
     {
         ApplicationUser addonUser = getAddOnUser();
         List<Project> allProjects = projectService.getAllProjects(addonUser).getReturnedValue();
@@ -294,7 +318,11 @@ public abstract class AbstractJiraPermissionScopeTest
     {
         ApplicationUser admin = userManager.getUserByKey(ADMIN);
         ProjectService.DeleteProjectValidationResult result = projectService.validateDeleteProject(admin.getDirectoryUser(), PROJECT_KEY);
-        projectService.deleteProject(admin, result);
+        
+        if(result.isValid())
+        {
+            projectService.deleteProject(admin, result);
+        }
     }
 
     protected void uninstallPlugin(Plugin plugin) throws IOException
